requirements.txt
sounddevice
numpy
soundfile
python-dotenv
openai-whisper
pyannote.audio